feature/EES-8 Create release summary versions.
