+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Services/ReleaseService.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
@@ -43,9 +44,9 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
             return await _context.Releases.FirstOrDefaultAsync(x => x.Id == id);
         }
         
-        public List<ContentSection> GetContent(ReleaseId id)
+        public async Task<List<ContentSection>> GetContentAsync(ReleaseId id)
         {
-            return _context.ContentSections.Where(section => section.ReleaseId == id).ToList();
+            return await _context.ContentSections.Where(section => section.ReleaseId == id).ToListAsync();
         }
         
         public List<Release> List()
@@ -70,22 +71,49 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
             return await ValidateReleaseSlugUniqueToPublication(createRelease.Slug, createRelease.PublicationId)
                 .OnSuccess(async () =>
                 {
-                    var order = OrderForNextReleaseOnPublication(createRelease.PublicationId);
-                    var content = TemplateFromRelease(createRelease.TemplateReleaseId);
+                    var releaseSummary = _mapper.Map<ReleaseSummaryVersion>(createRelease);
+                    releaseSummary.Created = DateTime.Now;
+                    
                     var release = _mapper.Map<Release>(createRelease);
-                    release.Content = content;
-                    release.Order = order;
-                    var saved = _context.Releases.Add(release);
+                    release.Content = await TemplateFromRelease(createRelease.TemplateReleaseId);
+                    release.Order = OrderForNextReleaseOnPublication(createRelease.PublicationId);
+                    release.ReleaseSummary = new ReleaseSummary
+                    {
+                        Versions = new List<ReleaseSummaryVersion>()
+                        {
+                            releaseSummary
+                        }
+                    };
+                    var saved =_context.Releases.Add(release);
                     await _context.SaveChangesAsync();
                     return await GetReleaseForIdAsync(saved.Entity.Id);
                 });
         }
 
         // TODO Authorisation will be required when users are introduced
+        public async Task<Either<ValidationResult, ReleaseViewModel>> EditReleaseSummaryAsyncOld(
+            ReleaseSummaryViewModel model)
+        {
+            var publication = await GetAsync(model.Id);
+            return await ValidateReleaseSlugUniqueToPublication(model.Slug, publication.Id, model.Id)
+                .OnSuccess(async () =>
+                {
+                    var release = await _context.Releases.FirstOrDefaultAsync(r => r.Id == model.Id);
+                    _context.Releases.Update(release);
+                    _mapper.Map(model, release);
+                    await _context.SaveChangesAsync();
+                    return await GetReleaseForIdAsync(model.Id);
+                });
+        }
+        
+        // TODO Authorisation will be required when users are introduced
         public async Task<ReleaseSummaryViewModel> GetReleaseSummaryAsync(ReleaseId releaseId)
         {
-            var release = await _context.Releases.FirstOrDefaultAsync(r => r.Id == releaseId);
-            return _mapper.Map<ReleaseSummaryViewModel>(release);
+            var release = await _context.Releases
+                .Where(r => r.Id == releaseId)
+                .HydrateReleaseForReleaseViewModel()
+                .FirstOrDefaultAsync();
+            return _mapper.Map<ReleaseSummaryViewModel>(release.ReleaseSummary);
         }
 
         // TODO Authorisation will be required when users are introduced
@@ -96,9 +124,17 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
             return await ValidateReleaseSlugUniqueToPublication(model.Slug, publication.Id, model.Id)
                 .OnSuccess(async () =>
                 {
-                    var release = await _context.Releases.FirstOrDefaultAsync(r => r.Id == model.Id);
-                    _context.Releases.Update(release);
+                    var release = await _context.Releases
+                        .Where(r => r.Id == model.Id)
+                        .HydrateReleaseForReleaseViewModel()
+                        .FirstOrDefaultAsync();
+                    
+                    var newSummary = _mapper.Map<ReleaseSummaryVersion>(model);
+                    newSummary.Created = DateTime.Now;
+                    newSummary.Summary = release.ReleaseSummary.Current.Summary;
                     _mapper.Map(model, release);
+                    release.ReleaseSummary.Versions.Add(newSummary);
+                    _context.Update(release);
                     await _context.SaveChangesAsync();
                     return await GetReleaseForIdAsync(model.Id);
                 });
@@ -142,11 +178,11 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
             return publication.Releases.Select(r => r.Order).DefaultIfEmpty().Max() + 1;
         }
 
-        private List<ContentSection> TemplateFromRelease(ReleaseId? releaseId)
+        private async Task<List<ContentSection>> TemplateFromRelease(ReleaseId? releaseId)
         {
             if (releaseId.HasValue)
             {
-                var templateContent = GetContent(releaseId.Value);
+                var templateContent = await GetContentAsync(releaseId.Value);
                 if (templateContent != null)
                 {
                     return templateContent.Select(c => new ContentSection
@@ -189,7 +225,8 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Services
                 .Include(r => r.Publication.Releases) // Back refs required to work out latest
                 .Include(r => r.Publication.Contact)
                 .Include(r => r.Type)
-                .Include(r => r.ReleaseSummary.Versions)
+                .Include(r => r.ReleaseSummary)
+                .ThenInclude(r => r.Versions)
                 .ThenInclude(v => v.Type);
         }
     }
